fix: 사용되는 숫자의 타입을 int에서 long으로 변경
- 큰 수가 들어올 수 있으니 범위가 큰 long으로 변경
- outputView에서는 출력할 데이터가 숫자인 것을 알 필요 없기 때문에 제네릭 메소드로 변경

+++ src/main/java/calculator/controller/Controller.java
@@ -25,7 +25,7 @@ public class Controller {
 
         Numbers numbers = numberExtractProgress(processedString, delimiters);
 
-        int result = calculateProgress(numbers);
+        long result = calculateProgress(numbers);
 
         OutputView.printResultMessage(result);
     }
@@ -38,7 +38,7 @@ public class Controller {
         return numberExtractor.extractNumber(inputString, delimiters);
     }
 
-    private int calculateProgress(Numbers numbers) {
+    private long calculateProgress(Numbers numbers) {
         return calculator.calculate(numbers);
     }
